refactor: move ClusterManager to cluster/manager.rs
